+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0.0 - 2016-10-24
+
+* Add : `little-loader` lib to fetch Google Map API
+* Add : `query-string` lib to transform params to query string
+* Update : `package.json` : description, author, dependencies
+* Update : `.babelrc`
+* Update : `.eslintrc`
+* Update : `src/index.js`
+
 ## 1.0.1 - 2015-09-11
 
 * Add : libraryTarget into Webpack config